#0: move groupnorm to ttnn, remove sweeps on broken groupnorm version
