--- tests/tt_eager/python_api_testing/sweep_tests/test_configs/ci_sweep_tests_broken/wormhole/pytorch_groupnorm_test.yaml
@@ -1,50 +1,0 @@
----
-test-list:
-  - groupnorm:
-      shape:
-        start-shape: [1, 1, 32, 32]
-        end-shape: [12, 24, 512, 512]
-        interval: [1, 1, 32, 32]
-        num-shapes: 3
-        num-samples: 64
-        args-sampling-strategy: "all"
-        method: groupnorm
-      datagen:
-        function: gen_rand
-        args:
-          low: -10
-          high: 10
-      comparison:
-        function: comp_pcc
-      args-gen: gen_dtype_layout_device
-      sanitize-args: False
-      output-file: groupnorm_sweep.csv
-      args:
-        data-layout: ["TILE"]
-        data-type: ["BFLOAT16", "BFLOAT8_B"]
-        buffer-type: ["DRAM", "L1", "SYSTEM_MEMORY"]
-        out-buffer-type: ["DRAM"]
-  - groupnorm:
-      shape:
-        start-shape: [1, 1, 2, 2]
-        end-shape: [12, 24, 512, 512]
-        interval: [1, 1, 2, 2]
-        num-shapes: 3
-        num-samples: 64
-        args-sampling-strategy: "all"
-        method: groupnorm
-      datagen:
-        function: gen_rand
-        args:
-          low: -10
-          high: 10
-      comparison:
-        function: comp_pcc
-      args-gen: gen_dtype_layout_device
-      sanitize-args: False
-      output-file: groupnorm_sweep.csv
-      args:
-        data-layout: ["ROW_MAJOR"]
-        data-type: ["BFLOAT16", "BFLOAT8_B"]
-        buffer-type: ["DRAM", "L1", "SYSTEM_MEMORY"]
-        out-buffer-type: ["DRAM"]
