--- tests/tt_eager/python_api_testing/sweep_tests/test_configs/ci_sweep_tests_working/grayskull/pytorch_groupnorm_noweights_test.yaml
@@ -1,26 +1,0 @@
----
-test-list:
-  - groupnorm-noweights:
-      shape:
-        start-shape: [1, 1, 32, 32]
-        end-shape: [6, 12, 256, 256]
-        interval: [1, 1, 32, 32]
-        num-shapes: 1
-        num-samples: 64
-        args-sampling-strategy: "all"
-      env:
-        # TT_PCI_DMA_BUF_SIZE: "1048576"
-      datagen:
-        function: gen_rand
-        args:
-          low: -10
-          high: 10
-      comparison:
-        function: comp_pcc
-      args-gen: gen_dtype_layout_device
-      output-file: groupnorm_noweights_sweep.csv
-      args:
-        data-layout: ["TILE"]
-        data-type: ["BFLOAT16"]
-        buffer-type: ["DRAM", "L1", "SYSTEM_MEMORY"]
-        out-buffer-type: ["DRAM", "L1"]
