+++ tests/ttnn/profiling/reference.txt
@@ -172,8 +172,6 @@ tt_lib.tensor.global_min,40,18.0,19.73,20.28,19.23
 tt_lib.tensor.global_sum,40,10.65,11.23,11.92,10.92
 tt_lib.tensor.glu_dim_2,40,2.97,3.59,3.6,3.41
 tt_lib.tensor.glu_dim_3,40,2.72,2.97,3.32,3.03
-tt_lib.tensor.groupnorm,40,7.53,8.57,9.35,8.51
-tt_lib.tensor.groupnorm_no_weights,40,5.93,7.46,8.25,7.3
 ttnn.gt,40,0.9,0.93,1.23,0.99
 tt_lib.tensor.gt_bw,40,0.98,1.07,1.36,1.11
 ttnn.ge,40,0.89,0.94,1.21,0.99
