+++ tests/ttnn/python_api_testing/sweep_tests/op_map.py
@@ -537,10 +537,6 @@ op_map = {
         "tt_op": ttnn_ops.groupnorm_noweights,
         "pytorch_op": pytorch_ops.groupnorm_noweights,
     },
-    "ttnn-groupnorm": {
-        "tt_op": ttnn_ops.groupnorm,
-        "pytorch_op": ttnn_pytorch_ops.groupnorm,
-    },
     "ttnn-global-avg-pool2d": {
         "tt_op": ttnn_ops.global_avg_pool2d,
         "pytorch_op": pytorch_ops.global_avg_pool2d,
