--- tests/ttnn/python_api_testing/sweep_tests/test_configs/ci_sweep_tests_broken/wormhole/ttnn_groupnorm_noweights_test.yaml
@@ -1,28 +1,0 @@
----
-test-list:
-  - ttnn-groupnorm_noweights:
-      shape:
-        start-shape: [1, 1, 32, 32]
-        end-shape: [6, 12, 256, 256]
-        interval: [1, 1, 32, 32]
-        num-dims: [2, 3, 4]
-        num-shapes: 1
-        num-samples: 128
-        args-sampling-strategy: "all"
-      env:
-        # TT_PCI_DMA_BUF_SIZE: "1048576"
-      datagen:
-        function: gen_rand
-        args:
-          low: -100
-          high: 100
-      comparison:
-        function: comp_pcc
-      args-gen: gen_dtype_layout_device
-      sanitize-args: False
-      args:
-        data-layout: ["TILE"]
-        data-type: ["BFLOAT16"]
-        buffer-type: ["DRAM", "L1"]
-        out-buffer-type: ["DRAM", "L1", "SYSTEM_MEMORY"]
-      output-file: gorupnorm_noweights_sweep.csv
