--- tests/ttnn/python_api_testing/sweep_tests/test_configs/ci_sweep_tests_broken/wormhole/ttnn_groupnorm_test.yaml
@@ -1,29 +1,0 @@
----
-test-list:
-  - ttnn-groupnorm:
-      shape:
-        start-shape: [32, 64]
-        end-shape: [256, 512]
-        interval: [32, 32]
-        #num-dims: [2]
-        num-shapes: 3
-        num-samples: 128
-        args-sampling-strategy: "all"
-        method: ttnn-groupnorm
-      env:
-        # TT_PCI_DMA_BUF_SIZE: "1048576"
-      datagen:
-        function: gen_rand
-        args:
-          low: -100
-          high: 100
-      comparison:
-        function: comp_pcc
-      args-gen: gen_dtype_layout_device
-      sanitize-args: False
-      args:
-        data-layout: ["TILE"]
-        data-type: ["BFLOAT16"]
-        buffer-type: ["DRAM", "L1"]
-        out-buffer-type: ["DRAM", "L1", "SYSTEM_MEMORY"]
-      output-file: ttnn_groupnorm_sweep.csv
