+++ tests/ttnn/python_api_testing/sweep_tests/ttnn_ops.py
@@ -2207,27 +2207,6 @@ def groupnorm_noweights(
     return ttnn_tensor_to_torch(t1)
 
 
-def groupnorm(
-    x,
-    y,
-    z,
-    *args,
-    device,
-    dtype,
-    layout,
-    input_mem_config,
-    output_mem_config,
-    **kwargs,
-):
-    t0 = setup_ttnn_tensor(x, device, layout[0], input_mem_config[0], dtype[0])
-    t1 = setup_ttnn_tensor(y, device, layout[1], input_mem_config[1], dtype[1])
-    t2 = setup_ttnn_tensor(z, device, layout[2], input_mem_config[2], dtype[2])
-
-    t3 = ttnn.group_norm(t0, num_groups=1, weight=t1, bias=t2, memory_config=memory_config_to_ttnn(output_mem_config))
-
-    return ttnn_tensor_to_torch(t3)
-
-
 def eltwise_addcdiv(
     x,
     y,
