+++ tt_eager/tt_lib/csrc/operations/primary/module.hpp
@@ -8,7 +8,6 @@
 #include <pybind11/stl.h>
 
 #include "transformers/module.hpp"
-#include "tt_dnn/op_library/groupnorm/groupnorm_op.hpp"
 #include "tt_dnn/op_library/layernorm_distributed/layernorm_pre_allgather_op.hpp"
 #include "tt_dnn/op_library/layernorm_distributed/layernorm_post_allgather_op.hpp"
 #include "tt_dnn/op_library/moreh_adam/moreh_adam_op.hpp"
@@ -579,31 +578,6 @@ void py_module(py::module& m_primary) {
         py::arg("momentum_buffer_out_mem_config").noconvert() = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
         py::arg("compute_kernel_config").noconvert() = std::nullopt,
         "Performs a SGD operation.");
-
-    py::class_<GroupNormShardedMultiCoreProgramConfig>(m_primary, "GroupNormShardedMultiCoreProgramConfig")
-        .def(
-            py::init<CoreCoord, MathFidelity, DataType, DataType, bool>(),
-            py::kw_only(),
-            py::arg("compute_with_storage_grid_size"),
-            py::arg("math_fidelity").noconvert() = MathFidelity::HiFi4,
-            py::arg("im_data_format").noconvert() = DataType::BFLOAT16,
-            py::arg("out_data_format").noconvert() = DataType::BFLOAT16,
-            py::arg("inplace").noconvert() = false);
-
-    m_primary.def(
-        "groupnorm",
-        &groupnorm,
-        py::arg("input").noconvert(),
-        py::arg("num_groups").noconvert(),
-        py::arg("eps").noconvert(),
-        py::arg("gamma").noconvert() = std::nullopt,
-        py::arg("beta").noconvert() = std::nullopt,
-        py::arg("input_mask").noconvert() = std::nullopt,
-        py::arg("output_mem_config").noconvert() = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
-        py::arg("program_config").noconvert() = GroupNormShardedMultiCoreProgramConfig{},
-        R"doc(
-            Performs a groupnorm operation, returna a output tensor the same shape as input.
-        )doc");
 
     // moreh_groupnorm
     m_primary.def(
